Move the internal actions to a separate class.
